backend_prototype/requirements.txt
fastapi
google-generativeai
uvicorn